// project-files/src/server/types.ts
export interface IAutoDataRes {
    id: number;
    year: number;
    price: number;
    name: string;
    color: string;
    model: string;
    latitude: number;
    longitude: number;
};

export interface IAutoData {
    id: number;
    year: number;
    price: number;
    mark: string;
    color: string;
    model: string;
    latitude: number;
    longitude: number;
}